Fix chapters getting deleted when marking as unread from library (fixes #5755)
(cherry picked from commit 5ea8d0546e1eed8f799c3e49a50aa2f05d31bb1e)

+++ app/src/main/java/eu/kanade/tachiyomi/ui/library/LibraryPresenter.kt
@@ -710,7 +710,7 @@ class LibraryPresenter(
                 }
                 db.updateChaptersProgress(chapters).executeAsBlocking()
 
-                if (preferences.removeAfterMarkedAsRead()) {
+                if (read && preferences.removeAfterMarkedAsRead()) {
                     deleteChapters(manga, chapters)
                 }
             }
